+++ CHANGELOG.md
@@ -22,6 +22,7 @@ ChangeLog
 - BUGFIX: fix "t2a-supertonic" node for automatically adapting to changing sampling rates
 - BUGFIX: fix "t2t-subtitle" buffer handling for "import" mode
 - BUGFIX: fix "x2x-filter" node by not stumbling over undefined values
+- BUGFIX: fix "xio-websocket" node by closing all client connections on destruction
 - UPDATE: upgrade NPM dependencies and patches
 - CLEANUP: add missing type definitions for Kitten-TTS
 - CLEANUP: update year in all copyright messages